+++ src/screens/home/Home.js
@@ -7,26 +7,15 @@ import CreateEvent from './Admin/CreateEvent'
 import CurrentEvent from './CurrentEvent'
 
 const Home = ({ navigation }) => {
-    const [status, setStatus] = useState('');
-    const [empId, setEmpId] = useState('');
-    const [name, setName] = useState('');
-    const [email, setEmail] = useState('');
-    const [unit, setUnit] = useState('');
-    const [subunit, setSubunit] = useState('');
     const [ipAddress, setIpAddress] = useState('')
     const [permission, setPermission] = useState('')
     checkIpAddress().then(res => {
         setIpAddress(res)
     })
 
-    const handleSignOut = () => {
-        firebase.auth()
-            .signOut()
-            .then(() => {
-                navigation.replace('Welcome')
-            })
-            .catch(error => console.log(error.message))
-    }
+    useEffect(() => {
+        getCurrentEmployee()
+    })
 
     getCurrentEmployee = () => {
         firebase.firestore()
@@ -35,69 +24,15 @@ const Home = ({ navigation }) => {
             .get()
             .then(querySnapshot => {
                 querySnapshot.forEach(documentSnapshot => {
-                    getStatusEmployee(documentSnapshot.data()['status_id'])
-                    getSubunits(documentSnapshot.data()['subunit_id'])
-                    setEmpId(documentSnapshot.id)
-                    setEmail(documentSnapshot.data()['email'])
-                    setName(documentSnapshot.data()['full_name'])
                     setPermission(documentSnapshot.data()['permission_id'])
                 });
             });
     }
-
-    getStatusEmployee = (status_id) => {
-        const status = firebase.firestore()
-            .collection('status')
-            .doc(status_id.toString())
-            .onSnapshot(documentSnapshot => {
-                setStatus(documentSnapshot.data()['name'])
-                getUnits(documentSnapshot.id)
-            });
-        return () => status();
-    }
-
-    getSubunits = (subunit_id) => {
-        const subunit = firebase.firestore()
-            .collection('subunits')
-            .doc(subunit_id.toString())
-            .onSnapshot(documentSnapshot => {
-                setSubunit(documentSnapshot.data()['name'])
-            });
-        return () => subunit();
-    }
-
-    getUnits = (unit_id) => {
-        const subunit = firebase.firestore()
-            .collection('units')
-            .doc(unit_id.toString())
-            .onSnapshot(documentSnapshot => {
-                setUnit(documentSnapshot.data()['name'])
-            });
-        return () => subunit();
-    }
-
-    getCurrentEmployee()
-
-
-
     return (
         <View>
             <View>
                 <CreateEvent props={{ ipAddress: ipAddress, permission: permission }} />
             </View>
-            <View><Text>Ip Address: {ipAddress}</Text></View>
-            <View><Text>Employee Id: {empId}</Text></View>
-            <View><Text>Email: {email}</Text></View>
-            <View><Text>User: {name}</Text></View>
-            <View><Text>Status: {status}</Text></View>
-            <View><Text>Unit: {unit} | {subunit}</Text></View>
-            <Geolocation />
-            <TouchableOpacity onPress={() => { handleSignOut() }}>
-                <Text>Sign Out</Text>
-            </TouchableOpacity>
-            {/* <View>
-                <CurrentEvent />
-            </View> */}
         </View>
     )
 }
