+++ src/screens/index.js
@@ -3,3 +3,4 @@ export { default as Profile } from './home/Profile'
 export { default as SignIn } from './auth/SignIn'
 export { default as SignUp } from './auth/SignUp'
 export { default as Welcome } from './Welcome'
+export { default as CreateEvent } from './home/Admin/CreateEvent'
